+++ src/http/body.rs
@@ -34,12 +34,6 @@ impl Default for Body {
     }
 }
 
-impl Default for Body {
-    fn default() -> Body {
-        Body::empty()
-    }
-}
-
 impl Stream for Body {
     type Item = Chunk;
     type Error = ::Error;
@@ -114,6 +108,7 @@ impl From<&'static str> for Body {
 }
 
 impl From<Option<Body>> for Body {
+    #[inline]
     fn from (body: Option<Body>) -> Body {
         body.unwrap_or_default()
     }
