+++ src/http/request.rs
@@ -1,10 +1,11 @@
 use std::fmt;
 
 use header::Headers;
-use http::{Body, RequestHead};
+use http::{Body, MessageHead, RequestHead, RequestLine};
 use method::Method;
 use uri::{self, Uri};
 use version::HttpVersion;
+use std::net::SocketAddr;
 
 /// A client request to a remote server.
 pub struct Request<B = Body> {
@@ -14,6 +15,7 @@ pub struct Request<B = Body> {
     headers: Headers,
     body: Option<B>,
     is_proxy: bool,
+    remote_addr: Option<SocketAddr>,
 }
 
 impl<B> Request<B> {
@@ -27,13 +29,14 @@ impl<B> Request<B> {
             headers: Headers::new(),
             body: None,
             is_proxy: false,
+            remote_addr: None,
         }
     }
 
     /// Read the Request Uri.
     #[inline]
     pub fn uri(&self) -> &Uri { &self.uri }
-    
+
     /// Read the Request Version.
     #[inline]
     pub fn version(&self) -> HttpVersion { self.version }
@@ -48,8 +51,29 @@ impl<B> Request<B> {
 
     /// Read the Request body.
     #[inline]
-    pub fn body(&self) -> Option<&B> { self.body.as_ref() }
-    
+    pub fn body_ref(&self) -> Option<&B> { self.body.as_ref() }
+
+    /// The remote socket address of this request
+    ///
+    /// This is an `Option`, because some underlying transports may not have
+    /// a socket address, such as Unix Sockets.
+    ///
+    /// This field is not used for outgoing requests.
+    #[inline]
+    pub fn remote_addr(&self) -> Option<SocketAddr> { self.remote_addr }
+
+    /// The target path of this Request.
+    #[inline]
+    pub fn path(&self) -> &str {
+        self.uri.path()
+    }
+
+    /// The query string of this Request.
+    #[inline]
+    pub fn query(&self) -> Option<&str> {
+        self.uri.query()
+    }
+
     /// Set the Method of this request.
     #[inline]
     pub fn set_method(&mut self, method: Method) { self.method = method; }
@@ -78,14 +102,57 @@ impl<B> Request<B> {
     pub fn set_proxy(&mut self, is_proxy: bool) { self.is_proxy = is_proxy; }
 }
 
+impl Request<Body> {
+    /// Deconstruct this Request into its pieces.
+    ///
+    /// Modifying these pieces will have no effect on how hyper behaves.
+    #[inline]
+    pub fn deconstruct(self) -> (Method, Uri, HttpVersion, Headers, Body) {
+        (self.method, self.uri, self.version, self.headers, self.body.unwrap_or_default())
+    }
+
+    /// Take the Request body.
+    #[inline]
+    pub fn body(self) -> Body { self.body.unwrap_or_default() }
+}
+
 impl<B> fmt::Debug for Request<B> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_struct("Request")
             .field("method", &self.method)
             .field("uri", &self.uri)
             .field("version", &self.version)
+            .field("remote_addr", &self.remote_addr)
             .field("headers", &self.headers)
             .finish()
+    }
+}
+
+struct MaybeAddr<'a>(&'a Option<SocketAddr>);
+
+impl<'a> fmt::Display for MaybeAddr<'a> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match *self.0 {
+            Some(ref addr) => fmt::Display::fmt(addr, f),
+            None => f.write_str("None"),
+        }
+    }
+}
+
+/// Constructs a request using a received ResponseHead and optional body
+pub fn from_wire<B>(addr: Option<SocketAddr>, incoming: RequestHead, body: B) -> Request<B> {
+    let MessageHead { version, subject: RequestLine(method, uri), headers } = incoming;
+    debug!("Request::new: addr={}, req=\"{} {} {}\"", MaybeAddr(&addr), method, uri, version);
+    debug!("Request::new: headers={:?}", headers);
+
+    Request::<B> {
+        method: method,
+        uri: uri,
+        headers: headers,
+        version: version,
+        remote_addr: addr,
+        body: Some(body),
+        is_proxy: false,
     }
 }
