+++ src/lib.rs
@@ -39,6 +39,8 @@ pub use client::Client;
 pub use error::{Result, Error};
 pub use header::Headers;
 pub use http::{Body, Chunk};
+pub use http::request::Request;
+pub use http::response::Response;
 pub use method::Method::{self, Get, Head, Post, Put, Delete};
 pub use status::StatusCode::{self, Ok, BadRequest, NotFound};
 pub use server::Server;
